pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "knowledge-bot"
version = "0.1.0"
description = "A Telegram bot that captures content, processes it, and stores insights in an Obsidian vault"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Telegram bot
    "python-telegram-bot>=21.0",

    # Audio/Podcast processing
    "faster-whisper>=1.0.0",
    "yt-dlp>=2024.1.0",
    "feedparser>=6.0.0",

    # Article extraction
    "trafilatura>=1.6.0",
    "httpx>=0.27.0",

    # AI
    "anthropic>=0.40.0",
    "sentence-transformers>=3.0.0",

    # Storage
    "numpy>=1.26.0",

    # Scheduling
    "apscheduler>=3.10.0",

    # Configuration
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
knowledge-bot = "src.bot:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
